Renaming mcglib to mcg since the static library was called libmcglib.a which is a bit awkward. Also, the includes look nicer this way
