+++ libs/mcg/docs/semantic_map_spec.pdf
