+++ libs/mcg/include/mcg/Connection.h
