+++ libs/mcg/include/mcg/Entity.h
