+++ libs/mcg/include/mcg/Object.h
