+++ libs/mcg/src/core/Entity.cpp
@@ -1,4 +1,4 @@
-#include "mcglib/Entity.h"
+#include "mcg/Entity.h"
 using namespace std;
 using json = nlohmann::json;
