--- libs/mcglib/examples/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(mcglib_tutorial)
